Implement day9 solution

+++ day9/part1/exit.go
@@ -3,21 +3,21 @@ package main
 const exitOpCode = 99
 
 type exit struct {
-	pointer int
+	pointer int64
 }
 
 func (e *exit) OpCode() int {
 	return exitOpCode
 }
 
-func (e *exit) Pointer() int {
+func (e *exit) Pointer() int64 {
 	return e.pointer
 }
 
-func (e *exit) Offset() int {
+func (e *exit) Offset() int64 {
 	return 1
 }
 
-func (e *exit) Execute(memory []int, relativeOffset int) {
+func (e *exit) Execute(memory []int64, relativeOffset int64) {
 	return
 }
